style: create styled-components
