+++ .babelrc
@@ -4,7 +4,7 @@
     [
       "module-resolver",
       {
-        "root": ["./src"]
+        "root": ["./src", "./helpers"]
       }
     ],
     [
