+++ helpers/setupBrowserEnv.js
@@ -1,0 +1,2 @@
+import browserEnv from 'browser-env'
+browserEnv()
