+++ src/packages/styled-components/__tests__/renderStyleSheet.test.js
@@ -1,0 +1,11 @@
+import test from 'ava'
+
+import renderRule from '../renderStyleSheet'
+
+test('should return some classes', t => {
+  const rule = renderRule(() => ({
+    background: 'green'
+  }))
+
+  console.log(rule)
+})
