+++ src/packages/styled-components/renderStyleSheet.js
@@ -1,0 +1,15 @@
+import { h } from 'hyperapp' // eslint-disable-line
+import { createRenderer } from 'fela'
+import { render } from 'fela-dom'
+
+// intialize renderer
+const renderer = createRenderer()
+
+// setup stylesheet render
+const stylesheet = document.createElement('style')
+document.head.appendChild(stylesheet)
+render(renderer, stylesheet)
+
+// export renderRule, which attaches rules to the renerer instance
+const { renderRule } = renderer
+export default renderRule
